contract sign moved to PassportView
